add about and how to pages
