+++ frontend/src/pages/FlightPredictor.css
@@ -1,0 +1,12 @@
+.container {
+    margin: 30px;
+  }
+  
+  .input-field {
+    padding-top: 15px;
+  }
+  
+  .btn {
+    margin-top: 20px;
+    margin-bottom: 20px;
+  }
