+++ frontend/src/pages/HowToUse.css
@@ -1,0 +1,20 @@
+.how-to-use-container {
+    font-family: Arial, sans-serif;
+    margin-top: 20px;
+  }
+  
+  .how-to-use-container h1 {
+    font-size: 2.5rem;
+    margin-bottom: 1rem;
+  }
+  
+  .how-to-use-container ul {
+    list-style-type: disc;
+    margin-left: 20px;
+    line-height: 1.6;
+  }
+  
+  .how-to-use-container p {
+    line-height: 1.8;
+  }
+
